+++ src/main/java/io/lindx/sec/service/UserService.java
@@ -5,10 +5,10 @@ import java.util.List;
 import io.lindx.sec.models.User;
 
 public interface UserService {
-  User getUser(Integer id);
+  User getUser(Long id);
   User getByMail(String mail);
   void setUser(User user);
-  void setPassword(Integer id, String password);
+  void setPassword(Long id, String password);
   User getByName(String name);
   List<User> getAll();
 }
